+++ relay-dynamic-config/src/defaults.rs
@@ -192,14 +192,6 @@ pub fn hardcoded_span_metrics() -> Vec<(GroupKey, Vec<MetricSpec>, Vec<TagMappin
     let is_app_start = RuleCondition::glob("span.op", "app.start.*")
         & RuleCondition::eq("span.description", APP_START_ROOT_SPAN_DESCRIPTIONS);
 
-    // Metrics for common modules:
-    let is_common = is_app_start.clone()
-        | is_db.clone()
-        | is_resource.clone()
-        | is_http.clone()
-        | is_mobile.clone()
-        | is_interaction.clone();
-
     // Metrics for addon modules are only extracted if the feature flag is enabled:
     let is_addon = is_ai.clone() | is_queue_op.clone() | is_cache.clone();
 
@@ -211,7 +203,7 @@ pub fn hardcoded_span_metrics() -> Vec<(GroupKey, Vec<MetricSpec>, Vec<TagMappin
                     category: DataCategory::Span,
                     mri: "d:spans/exclusive_time@millisecond".into(),
                     field: Some("span.exclusive_time".into()),
-                    condition: Some(!is_addon.clone() & is_common.clone()),
+                    condition: Some(!is_addon.clone()),
                     tags: vec![],
                 },
                 MetricSpec {
@@ -286,7 +278,7 @@ pub fn hardcoded_span_metrics() -> Vec<(GroupKey, Vec<MetricSpec>, Vec<TagMappin
                     category: DataCategory::Span,
                     mri: "d:spans/duration@millisecond".into(),
                     field: Some("span.duration".into()),
-                    condition: Some(!is_addon.clone() & is_common.clone()),
+                    condition: Some(!is_addon.clone()),
                     tags: vec![],
                 },
                 MetricSpec {